PIM-3595: move base saver and remover in storage utils
