+++ src/Akeneo/Bundle/StorageUtilsBundle/Doctrine/Common/Remover/BaseRemover.php
@@ -1,62 +1,62 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Saver;
+namespace Akeneo\Bundle\StorageUtilsBundle\Doctrine\Common\Remover;
 
-use Akeneo\Component\StorageUtils\Saver\BulkSaverInterface;
+use Akeneo\Component\StorageUtils\Remover\BulkRemoverInterface;
+use Akeneo\Component\StorageUtils\Remover\RemoverInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\Common\Util\ClassUtils;
-use Akeneo\Component\StorageUtils\Saver\SaverInterface;
 
 /**
- * Base saver, declared as different services for different classes
+ * Base remover, declared as different services for different classes
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class BaseSaver implements SaverInterface, BulkSaverInterface
+class BaseRemover implements RemoverInterface, BulkRemoverInterface
 {
     /** @var ObjectManager */
     protected $objectManager;
 
-    /** @var BaseSavingOptionsResolver */
+    /** @var BaseRemovingOptionsResolver */
     protected $optionsResolver;
 
     /** @var string */
-    protected $savedClass;
+    protected $removedClass;
 
     /**
-     * @param ObjectManager             $objectManager
-     * @param BaseSavingOptionsResolver $optionsResolver
-     * @param string                    $savedClass
+     * @param ObjectManager               $objectManager
+     * @param BaseRemovingOptionsResolver $optionsResolver
+     * @param string                      $removedClass
      */
     public function __construct(
         ObjectManager $objectManager,
-        BaseSavingOptionsResolver $optionsResolver,
-        $savedClass
+        BaseRemovingOptionsResolver $optionsResolver,
+        $removedClass
     ) {
         $this->objectManager = $objectManager;
         $this->optionsResolver = $optionsResolver;
-        $this->savedClass = $savedClass;
+        $this->removedClass = $removedClass;
     }
 
     /**
      * {@inheritdoc}
      */
-    public function save($object, array $options = [])
+    public function remove($object, array $options = [])
     {
-        if (!$object instanceof $this->savedClass) {
+        if (!$object instanceof $this->removedClass) {
             throw new \InvalidArgumentException(
                 sprintf(
                     'Expects a "%s", "%s" provided.',
-                    $this->savedClass,
+                    $this->removedClass,
                     ClassUtils::getClass($object)
                 )
             );
         }
 
-        $options = $this->optionsResolver->resolveSaveOptions($options);
-        $this->objectManager->persist($object);
+        $options = $this->optionsResolver->resolveRemoveOptions($options);
+        $this->objectManager->remove($object);
 
         if (true === $options['flush'] && true === $options['flush_only_object']) {
             $this->objectManager->flush($object);
@@ -68,18 +68,18 @@ class BaseSaver implements SaverInterface, BulkSaverInterface
     /**
      * {@inheritdoc}
      */
-    public function saveAll(array $objects, array $options = [])
+    public function removeAll(array $objects, array $options = [])
     {
         if (empty($objects)) {
             return;
         }
 
-        $allOptions = $this->optionsResolver->resolveSaveAllOptions($options);
+        $allOptions = $this->optionsResolver->resolveRemoveAllOptions($options);
         $itemOptions = $allOptions;
         $itemOptions['flush'] = false;
 
         foreach ($objects as $object) {
-            $this->save($object, $itemOptions);
+            $this->remove($object, $itemOptions);
         }
 
         if (true === $allOptions['flush']) {
