+++ src/Akeneo/Bundle/StorageUtilsBundle/Doctrine/Common/Saver/BaseSaver.php
@@ -1,62 +1,62 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Remover;
+namespace Akeneo\Bundle\StorageUtilsBundle\Doctrine\Common\Saver;
 
-use Akeneo\Component\StorageUtils\Remover\BulkRemoverInterface;
-use Akeneo\Component\StorageUtils\Remover\RemoverInterface;
+use Akeneo\Component\StorageUtils\Saver\BulkSaverInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\Common\Util\ClassUtils;
+use Akeneo\Component\StorageUtils\Saver\SaverInterface;
 
 /**
- * Base remover, declared as different services for different classes
+ * Base saver, declared as different services for different classes
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class BaseRemover implements RemoverInterface, BulkRemoverInterface
+class BaseSaver implements SaverInterface, BulkSaverInterface
 {
     /** @var ObjectManager */
     protected $objectManager;
 
-    /** @var BaseRemovingOptionsResolver */
+    /** @var BaseSavingOptionsResolver */
     protected $optionsResolver;
 
     /** @var string */
-    protected $removedClass;
+    protected $savedClass;
 
     /**
-     * @param ObjectManager               $objectManager
-     * @param BaseRemovingOptionsResolver $optionsResolver
-     * @param string                      $removedClass
+     * @param ObjectManager             $objectManager
+     * @param BaseSavingOptionsResolver $optionsResolver
+     * @param string                    $savedClass
      */
     public function __construct(
         ObjectManager $objectManager,
-        BaseRemovingOptionsResolver $optionsResolver,
-        $removedClass
+        BaseSavingOptionsResolver $optionsResolver,
+        $savedClass
     ) {
         $this->objectManager = $objectManager;
         $this->optionsResolver = $optionsResolver;
-        $this->removedClass = $removedClass;
+        $this->savedClass = $savedClass;
     }
 
     /**
      * {@inheritdoc}
      */
-    public function remove($object, array $options = [])
+    public function save($object, array $options = [])
     {
-        if (!$object instanceof $this->removedClass) {
+        if (!$object instanceof $this->savedClass) {
             throw new \InvalidArgumentException(
                 sprintf(
                     'Expects a "%s", "%s" provided.',
-                    $this->removedClass,
+                    $this->savedClass,
                     ClassUtils::getClass($object)
                 )
             );
         }
 
-        $options = $this->optionsResolver->resolveRemoveOptions($options);
-        $this->objectManager->remove($object);
+        $options = $this->optionsResolver->resolveSaveOptions($options);
+        $this->objectManager->persist($object);
 
         if (true === $options['flush'] && true === $options['flush_only_object']) {
             $this->objectManager->flush($object);
@@ -68,18 +68,18 @@ class BaseRemover implements RemoverInterface, BulkRemoverInterface
     /**
      * {@inheritdoc}
      */
-    public function removeAll(array $objects, array $options = [])
+    public function saveAll(array $objects, array $options = [])
     {
         if (empty($objects)) {
             return;
         }
 
-        $allOptions = $this->optionsResolver->resolveRemoveAllOptions($options);
+        $allOptions = $this->optionsResolver->resolveSaveAllOptions($options);
         $itemOptions = $allOptions;
         $itemOptions['flush'] = false;
 
         foreach ($objects as $object) {
-            $this->remove($object, $itemOptions);
+            $this->save($object, $itemOptions);
         }
 
         if (true === $allOptions['flush']) {
